refactor(server): rename Server -> LegacyServer (verbatim)
git mv server.ts legacyServer.ts + class rename. Zero body changes.
Importers updated to `LegacyServer as Server` so behavior is identical
and the rename shows as a true file rename in history.
